requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.26.137
elasticsearch[async]==8.11.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
python-jose==3.3.0
prometheus-client==0.19.0
aiofiles==23.2.1
jinja2==3.1.2
openai==1.3.0
